split trait+impl
