+++ lib/client/src/account_fetchers.rs
@@ -1,3 +1,5 @@
+// TODO move to feeds
+
 use std::collections::HashMap;
 use std::sync::Arc;
 use std::sync::Mutex;
@@ -14,22 +16,8 @@ use solana_sdk::account::{AccountSharedData, ReadableAccount};
 use solana_sdk::pubkey::Pubkey;
 
 use mango_v4::state::MangoAccountValue;
+use crate::AccountFetcher;
 
-#[async_trait::async_trait]
-pub trait AccountFetcher: Sync + Send {
-    async fn fetch_raw_account(&self, address: &Pubkey) -> anyhow::Result<AccountSharedData>;
-    async fn fetch_raw_account_lookup_table(
-        &self,
-        address: &Pubkey,
-    ) -> anyhow::Result<AccountSharedData> {
-        self.fetch_raw_account(address).await
-    }
-    async fn fetch_program_accounts(
-        &self,
-        program: &Pubkey,
-        discriminator: [u8; 8],
-    ) -> anyhow::Result<Vec<(Pubkey, AccountSharedData)>>;
-}
 
 // Can't be in the trait, since then it would no longer be object-safe...
 pub async fn account_fetcher_fetch_anchor_account<T: AccountDeserialize>(
@@ -139,7 +127,7 @@ struct AccountCache {
 
     account_jobs: CoalescedAsyncJob<Pubkey, anyhow::Result<AccountSharedData>>,
     program_accounts_jobs:
-        CoalescedAsyncJob<(Pubkey, [u8; 8]), anyhow::Result<Vec<(Pubkey, AccountSharedData)>>>,
+    CoalescedAsyncJob<(Pubkey, [u8; 8]), anyhow::Result<Vec<(Pubkey, AccountSharedData)>>>,
 }
 
 impl AccountCache {
